Separate out keys tests. Add taproot test to sanity-check wallet-generated addresses and descriptors.
